+++ excalidraw/src/types.ts
@@ -1,6 +1,0 @@
-export type VSCodeApi = {
-  postMessage(message: any): void;
-  getState(): any;
-  setState(state: any): void;
-};
-