+++ src/fs.ts
@@ -1,4 +1,5 @@
-import { TextDecoder, TextEncoder } from "util";
+import path = require("path");
+import { TextDecoder } from "util";
 import * as vscode from "vscode";
 
 /**
@@ -17,60 +18,48 @@ export function editDoc(
   return vscode.workspace.applyEdit(edit);
 }
 
-export async function loadWebviewContent(
-  uri: vscode.Uri,
-  documentUri: vscode.Uri,
-  documentType: "application/json" | "image/svg" = "application/json"
+export async function getWebviewContent(
+  templateUri: vscode.Uri,
+  documentWebviewUri: vscode.Uri,
+  config: {
+    theme: "light" | "dark",
+    viewModeEnabled?: boolean;
+    gridModeEnable?: boolean;
+    zenModeEnabled?: boolean;
+  }
 ): Promise<string> {
-  let raw = await vscode.workspace.fs.readFile(uri);
-  // if (initialData.readOnly)
-  //   content = new TextDecoder().decode(raw).replace(
-  //     /(<style>)(<\/style>)/,
-  //     `$1
-  // 		.Island {
-  // 			display: none !important;
-  // 		}
-  // 		$2`
-  //   );
+  const extName = path.extname(documentWebviewUri.fsPath);
+  const jsonConfig = JSON.stringify(config);
+  const bufferConfig = Buffer.from(jsonConfig, "utf-8");
+  const base64Config = bufferConfig.toString("base64");
 
+  let TemplateRaw = await vscode.workspace.fs.readFile(templateUri);
   const html = new TextDecoder()
-    .decode(raw)
-    .replace("{{excalidraw-document-uri}}", documentUri.toString(true))
-	.replace("{{excalidraw-document-type}}", documentType)
+    .decode(TemplateRaw)
+    .replace("{{excalidraw-document-uri}}", documentWebviewUri.toString(true))
+    .replace(
+      "{{data-document-type}}",
+      extName === ".svg" ? "image/svg+xml" : "application/json"
+    )
+    .replace("{{data-config}}", base64Config)
+    .replace("{{data-theme}}", config.theme);
+
   return html;
 }
 
-// function getExportFilename(
-//       document: vscode.TextDocument,
-//       extension: string
-//     ): Thenable<vscode.Uri | undefined> {
-//       const dirname = vscode.Uri.parse() path.dirname(document.uri.fsPath);
-//       const basename = path.basename(
-//         document.uri.fsPath,
-//         path.extname(document.uri.fsPath)
-//       );
-//       const globs: any = vscode.workspace
-//         .getConfiguration("excalidraw.export")
-//         .get("globs");
-//       const worskspaceFolder = vscode.workspace.workspaceFolders?.[0];
-//       for (const [glob, outputDir] of Object.entries(globs)) {
-//         if (minimatch(document.uri.fsPath, glob))
-//           return new Promise((resolve) => {
-//             if (worskspaceFolder === undefined || typeof outputDir != "string")
-//               resolve(undefined);
-//             else {
-//               const outputPath = path.join(
-//                 worskspaceFolder.uri.fsPath,
-//                 outputDir,
-//                 `${basename}.${extension}`
-//               );
-//               resolve(vscode.Uri.parse(outputPath));
-//             }
-//           });
-//       }
-//       const filePath = path.join(dirname, `${basename}.${extension}`);
-//       return vscode.window.showSaveDialog({
-//         defaultUri: vscode.Uri.file(filePath),
-//         filters: { Images: [extension] },
-//       });
-//     }
+export function getExportFilename(
+  document: vscode.TextDocument,
+  extension: string
+): Thenable<vscode.Uri | undefined> {
+  const dirname = path.dirname(document.uri.fsPath);
+  const basename = path.basename(
+    document.uri.fsPath,
+    path.extname(document.uri.fsPath)
+  );
+
+  const filePath = path.join(dirname, `${basename}.${extension}`);
+  return vscode.window.showSaveDialog({
+    defaultUri: vscode.Uri.file(filePath),
+    filters: { Images: [extension] },
+  });
+}